Move resources to testclasses in initializers
